+++ src/Sema/Type/TypeInfo.cpp
@@ -231,7 +231,7 @@ uint32_t TypeInfo::hash() const
         case TypeInfoKind::Array:
             h = Math::hashCombine(h, payloadArray_.typeRef.get());
             for (const auto dim : payloadArray_.dims)
-                h = Math::hashCombine(h, dim);
+                h = Math::hashCombine(h, static_cast<uint32_t>(dim));
             return h;
 
         default:
@@ -679,7 +679,7 @@ TypeInfo TypeInfo::makeSlice(TypeRef pointeeTypeRef, TypeInfoFlags flags)
     return ti;
 }
 
-TypeInfo TypeInfo::makeArray(const std::vector<uint64_t>& dims, TypeRef elementTypeRef, TypeInfoFlags flags)
+TypeInfo TypeInfo::makeArray(const std::vector<int64_t>& dims, TypeRef elementTypeRef, TypeInfoFlags flags)
 {
     TypeInfo ti{TypeInfoKind::Array, flags};
     std::construct_at(&ti.payloadArray_.dims, dims);
